refactor: unify the interface for both TESTICP and ICRC1 (#15)
This update further simplifies the design of the token faucet. Earlier,
we merged the frontend and backend into one canister, reducing the
number of canisters from four to two (one for TESTICP and another for
TICRC1). Now we merge the frontend to support both tokens, so that now
we have one canister, simplifying code, documentation, and
discoverability.