fix: correctly set canister ID in frontend (#22)
The `npm run build` command in the frontend wasn't properly setting up
the canister ID of the faucet. This wasn't a problem when building
locally, as vite was reading that configuration from the local `.env`
file. But, when building from our release script on a clean VM, without
the `.env` file being present, the canister ID wouldn't be set and the
frontend would be broken.